x/tollroad/types/messages_user_vault_test.go
package types

import (
	"testing"

	"github.com/b9lab/toll-road/testutil/sample"
	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgCreateUserVault_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgCreateUserVault
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgCreateUserVault{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgCreateUserVault{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}

func TestMsgUpdateUserVault_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgUpdateUserVault
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgUpdateUserVault{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgUpdateUserVault{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}

func TestMsgDeleteUserVault_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgDeleteUserVault
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgDeleteUserVault{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgDeleteUserVault{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
